backend/requirements.txt
amqp==5.3.1
aniso8601==9.0.1
bcrypt==4.2.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
